move main func to 1_snap_equal_test.cpp
