+++ source/module_orbital/test/1_snap_equal_test.cpp
@@ -71,3 +71,16 @@ TEST_F(test_orb, equal_test)
 	}
 }
 
+int main(int argc, char** argv)
+{
+#ifdef __MPI
+    MPI_Init(&argc,&argv);
+#endif
+
+    testing::InitGoogleTest(&argc, argv);
+    int result = RUN_ALL_TESTS();
+
+#ifdef __MPI
+    MPI_Finalize();
+#endif
+}
